+++ WallManager.h
@@ -11,7 +11,7 @@ public:
 	void AddWall(Wall* newWall) { _wallV.push_back(newWall); }
 	void EraseWall(Wall* targetWall);
 
-	bool CheckCross(int startX, int startY, int endX, int endY);
-	bool CheckCross(int startX, int startY, int endX, int endY, int idx);
+	bool CheckCross(float startX, float startY, float endX, float endY);
+	bool CheckCross(float startX, float startY, float endX, float endY, int idx);
 };
